Backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.13.1
asyncpg==0.29.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
websockets==12.0
redis==5.0.1
celery==5.3.4
faker==20.1.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
langchain==0.1.0
langgraph==0.0.20
langchain-openai==0.0.2
langchain-community==0.0.12
langchain-tavily==0.1.0
openai==1.6.1
supabase==2.0.2
mcp==0.3.0
python-dotenv==1.0.0
typer==0.9.0
rich==13.7.0
wikipedia==1.4.0 